refactor(capsule)!: extract wasm-clean astrid-capsule-types crate
Move the engine-agnostic capsule types (manifest, CapsuleId, fuel/memory
ledgers, runtime/HTTP limits) out of the Wasmtime engine crate into a new
wasmtime-free, tokio-free astrid-capsule-types crate so any capsule engine
can share the vocabulary the kernel routes and gates on. astrid-capsule
re-exports every moved item at its original path, so workspace consumers
compile unchanged. The manifest security surface (EnvScope
skip_deserializing) and its parser-isolation tests move intact.

BREAKING CHANGE: the tokio-Semaphore constructors on CapsuleRuntimeLimits
(blocking_semaphore/io_semaphore) are now methods of the
CapsuleRuntimeLimitsExt extension trait in astrid-capsule; external callers
must bring the trait into scope.

Closes #1133
