+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Changelog tracking starts with 0.2.0. Prior versions were not tracked.
 
 - **The kernel's pure-semantics crates now compile for `wasm32-unknown-unknown`, groundwork for a browser host profile.** `astrid-types`, `astrid-core`, `astrid-crypto`, `astrid-capabilities`, `astrid-audit`, `astrid-events`, `astrid-config`, and `astrid-approval` build for the wasm target with no behaviour change on native. Two structural blockers were cleared. First, the workspace-level `tokio` dependency pinned an additive base feature set (`net`, `signal`, `rt-multi-thread`) onto every `workspace = true` consumer — including leaf crates that only asked for `sync`/`time` — dragging in `mio` (which `compile_error!`s on wasm) and the multi-threaded runtime (which needs OS threads). The base is trimmed to the portable subset (`sync`, `macros`, `time`, `rt`, `io-util`); the native-only features are now declared locally by the crates that actually use them (uplink/cli/kernel/gateway/capsule for `net`, daemon/kernel for `signal`, daemon/emit/cli/capsule-install/capsule/hooks/kernel/gateway for `rt-multi-thread`). Second, `getrandom` (reached via `uuid`/`rand` at major 0.4 and `surrealkv` at major 0.3) refuses to build for wasm without a backend selection: the `wasm_js` feature is now enabled through `[target.'cfg(target_family = "wasm")'.dependencies]` entries on the crates owning those edges, paired with the build-time `--cfg getrandom_backend="wasm_js"` that `scripts/check-wasm-portability.sh` (a new `cargo check`-only regression guard, wired into CI) sets. The audit store's `block_in_place` fast path (native-only) is cfg-gated to its scoped-thread fallback on wasm. No public API changes; native `build`/`test`/`clippy`/`fmt` are unaffected.
 
+- **Engine-agnostic capsule types extracted into a new `astrid-capsule-types` crate.** The pure data the kernel routes and gates on — the capsule manifest (`CapsuleManifest` and its validation/error types), the capsule id (`CapsuleId`), the per-principal fuel and memory peak ledgers (`FuelLedger`/`FuelRateLimiter`, `MemoryLedger`), and the resource/HTTP limit types (`CapsuleRuntimeLimits`, `HttpLimits`) — lived inside `astrid-capsule`, the crate that owns the Wasmtime execution engine. Because those types were trapped in the engine crate, every component that needs only the *data* — the kernel included — transitively depended on Wasmtime, and a second capsule engine profile could not share the vocabulary without dragging in the engine it replaces. They now live in a `wasmtime`-free, `tokio`-free `astrid-capsule-types` crate that every engine can depend on; `astrid-capsule` re-exports each moved item at its exact original path, so the kernel and all other consumers compile unchanged. Behaviour that genuinely touches the engine stays behind: the tokio-`Semaphore` constructors on `CapsuleRuntimeLimits` become the `CapsuleRuntimeLimitsExt` extension trait, and the per-Store `StoreMemoryMeter` (`wasmtime::ResourceLimiter`) keeps feeding the shared `MemoryLedger`. A pure move for every consumer that uses the re-exported paths — no behaviour changes, and the manifest's untrusted-input `skip_deserializing` protections and parser-isolation tests move intact. One API-shape caveat for external consumers of `astrid-capsule`: the semaphore constructors are now trait methods, so `CapsuleRuntimeLimitsExt` must be in scope to call `blocking_semaphore()`/`io_semaphore()`. Closes #1133.
+
 ### Fixed
 
 - **`astrid stop`/`astrid restart` no longer report success while a daemon wedges mid-shutdown and leaks the state-db lock.** The graceful stop path treated the daemon's shutdown *acknowledgement* as proof it had exited: it printed "Astrid daemon stopped" and deleted the recorded PID file the moment the daemon ACK'd the `Shutdown` request — before the process actually exited and released the singleton/state-db lock. A daemon that then wedged during its own shutdown sequence (a capsule run-loop that would not cancel, a hung audit-DB flush) kept the lock forever, but the PID file — the only handle `restart`'s orphan-killer needs — was already gone, so `restart` skipped the kill and the next `start` died on the held lock with a raw "Database … is already locked by another process". Stop now captures the recorded PID *before* requesting shutdown, and after the ACK waits for that process to actually exit; if it is still alive past the grace window it escalates through the same identity-gated `SIGTERM`→`SIGKILL` path used for an unreachable orphan. Runtime files (socket, readiness, PID) are removed only once the daemon is confirmed gone — when a kill cannot confirm exit (`StillAlive`/`Unverified`) they are kept so `start`/`restart` still find the recorded PID and print an actionable message instead of failing on the lock. The unreachable-daemon path is also hardened to sweep a wedged process whose socket has already vanished. Closes #1120.
