+++ Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "crates/astrid-build",
     "crates/astrid-capsule",
     "crates/astrid-capsule-install",
+    "crates/astrid-capsule-types",
     "crates/astrid-cli",
     "crates/astrid-config",
     "crates/astrid-daemon",
@@ -44,6 +45,7 @@ astrid-capabilities = { path = "crates/astrid-capabilities", version = "0.9.2" }
 astrid-build = { path = "crates/astrid-build", version = "0.9.2" }
 astrid-capsule = { path = "crates/astrid-capsule", version = "0.9.2" }
 astrid-capsule-install = { path = "crates/astrid-capsule-install", version = "0.9.2" }
+astrid-capsule-types = { path = "crates/astrid-capsule-types", version = "0.9.2" }
 astrid-config = { path = "crates/astrid-config", version = "0.9.2" }
 astrid-daemon = { path = "crates/astrid-daemon", version = "0.9.2" }
 astrid-core = { path = "crates/astrid-core", version = "0.9.2" }
