+++ crates/astrid-capsule-types/src/error.rs
