+++ crates/astrid-capsule-types/src/fuel_ledger.rs
