+++ crates/astrid-capsule-types/src/manifest/capabilities.rs
