+++ crates/astrid-capsule-types/src/manifest/topics.rs
